feat(jans-auth-server): added new stat response service with test
https://github.com/JanssenProject/jans/issues/317
